Compound interest fix

+++ contracts/lending-pool/Config.sol
@@ -94,10 +94,10 @@ library Config {
         return _rate.rayMul(timeDelta).add(WadRayMath.ray());
     }
 
-    function calculateCompoundedInterest(uint256 _rate, uint40 _lastUpdateTimestamp)
+    function calculateCompoundedInterest(uint256 _rate, uint _lastUpdateTimestamp)
     internal view returns (uint256)
     {
-        uint256 timeDifference = block.timestamp.sub(uint256(_lastUpdateTimestamp));
+        uint256 timeDifference = block.timestamp.sub(_lastUpdateTimestamp);
         uint256 ratePerSecond = _rate.div(SECONDS_PER_YEAR);
         return ratePerSecond.add(WadRayMath.ray()).rayPow(timeDifference);
     }
